Added pull down resistor for RPI_PIN18 (2.2k) and RPI_PIN16 (2.2k NF).
These pins are connected directly to the CM4/5 module's Ethernet PHY and appear to be used as bootstrap configuration pins, although such information is not provided in the CM4/5 documentation.

Due to the connected unconfigured FPGA with internal pull ups, the levels of these lines change and the Ethernet phy no longer detected.

It seems that the level of the RPI_PIN18 line for CM4/5 must be low, and also the level of the RPI_PIN16 line for CM4 must be low (not critical?) and high for CM5.